增加专业的“我喜欢”功能 新增添加我喜欢接口 新增取消我喜欢接口 更新查询专业返回字段添加我喜欢的数量（likeCount）

+++ src/main/java/com/graduationprojectordermanagementsystem/interceptor/JwtAuthenticationInterceptor.java
@@ -71,7 +71,7 @@ public class JwtAuthenticationInterceptor implements HandlerInterceptor {
             //5.提取用户信息，存入上下文中
             String username = claims.getSubject();
             String role = claims.get("role", String.class);//获取角色
-            Integer userId = claims.get("userId", Integer.class);
+            Long userId = claims.get("userId", Long.class);
             if (username == null && role == null){
                 setUnauthorizedResponse(response, "Token中缺少必要信息");
                 return false;
