+++ src/main/java/com/graduationprojectordermanagementsystem/util/UserContext.java
@@ -7,7 +7,7 @@ public class UserContext {
 
     private static final ThreadLocal<String> usernameHolder = new ThreadLocal<>();
     private static final ThreadLocal<String> roleHolder = new ThreadLocal<>();
-    private static final ThreadLocal<Integer> userIdHolder = new ThreadLocal<>();
+    private static final ThreadLocal<Long> userIdHolder = new ThreadLocal<>();
 
     // 设置用户名
     public static void setUsername(String username) {
@@ -30,10 +30,10 @@ public class UserContext {
     }
 
     // 设置用户id
-    public static void setUserId(Integer userId) {userIdHolder.set(userId);}
+    public static void setUserId(Long userId) {userIdHolder.set(userId);}
 
     // 获取用户id
-    public static Integer getUserId() {return userIdHolder.get();}
+    public static Long getUserId() {return userIdHolder.get();}
 
     // 清除用户名
     public static void clear() {
